Move the new tests to `test_quadrupole.py` and update `CHANGELOG.md`
